requirements.txt
haystack-ai
transformers
streamlit
sentence-transformers